Se actualiza entity con getter y setters y se prueba con postaman ok

+++ src/main/java/com/pabloisla/demo/repositorios/EmpresaRepository.java
@@ -4,6 +4,6 @@ import com.pabloisla.demo.modelos.Empresa;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmpresaRepository extends JpaRepository<Empresa, Long> {
+public interface EmpresaRepository extends JpaRepository<Empresa, Integer> {
 
 }
